% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT dataset for fish}
\format{
## `who`
A data frame with 144 rows and 6 columns:
\describe{
  \item{LENGTH}{Length of fish}
  \item{WEIGHT}{Weight of fish}

  ...
}
}
\source{
<https://www.amazon.com/Statistics-Engineering-Sciences-William-Mendenhall/dp/1498728855>
}
\usage{
ddt
}
\description{
Fish measurements
}
\keyword{datasets}
