% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Data}
\format{
## `fire`

\describe{
  \item{x, Distance from fire station}{y, Fire damage in thousands of dollars}
  ...
}
}
\usage{
fire
}
\description{
Has two quantitative variables: x and y, describing distance from fire station
and fire damage respectively
}
\keyword{datasets}
