% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{iter}

\item{x}{x}

\item{fun}{fun}

\item{alpha}{alpha}

\item{cx}{cx}

\item{...}{s}
}
\description{
myboot2
}
