% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum Likelihood Function}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{log likelihood function}

\item{x}{values}

\item{param}{parameters}

\item{...}{ellipses argument}
}
\value{
plot
}
\description{
For repeated sampling from the same distribution
}
