% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{MySample}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{number of trials}

\item{iter}{number of iterations}

\item{time}{time, in seconds, between each iteration}
}
\value{
none, barplots printed to console
}
\description{
MySample
}
\examples{
\dontrun{mysample(n=1000, iter=30, time=1)}

}
