% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Airline Overbooking Function}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{The number of seats in the flight}

\item{gamma}{The probability of overbooking}

\item{p}{The probabiility that a passenger will show up}
}
\value{
A list containing nd, nc, n, p, and gamma.
}
\description{
Calculates the optimal number of seats to sell given the number of seats,
probability of overbooking, and the probability that a passenger will show up.
The problem can be modeled using a discrete (binomial) distribution or a
continuous (normal) distribution.
}
\details{
To calculate n for the discrete distribution (binomial), the function creates
a sequence of values from N to round(N+N/10), separated by 1. These values are
used to create a vector of (1-gamma)th quantiles using different "trials" (number
of tickets). This vector is subtracted by N so that the index of the minimum value, which
is 0, may be found. This index gives n when plugged into the sequence.

The same applies for the continuous distribution (normal), except it is the inverse cdf of the
normal distribution. The sequences are separated by small, fractional values instead.

The objective functions of the discrete and continuous are calculated the same way, except with
1-gamma-pbinom(...) and 1-gamma-pnorm(...)
}
