% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterhist.R
\name{scatterhist}
\alias{scatterhist}
\title{Plot Function}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{x variable}

\item{y}{y variable}

\item{xlab}{label for x variable}

\item{ylab}{label for y variable}
}
\value{
a scatter histogram
}
\description{
Creates a scatter-histogram plot with two variables, LENGTH and WEIGHT.
}
\examples{
with(ddt, scatterhist(LENGTH,WEIGHT, xlab="LENGTH"))
}
