% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage dataset}
\format{
## `who`
A data frame with 15 rows and 2 columns:
\describe{
  \item{DISTANCE}{Distance, in miles}
  \item{DAMAGE}{Damage, in tens of thousands of dollars}

  ...
}
}
\source{
<https://www.amazon.com/Statistics-Engineering-Sciences-William-Mendenhall/dp/1498728855>
}
\usage{
fire
}
\description{
Distance from fire station and damage in tens of thousands of dollars. Sample
size: 15
}
\keyword{datasets}
