% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{myclt}
\usage{
myclt(n, iter)
}
\arguments{
\item{n}{n}

\item{iter}{iter}
}
\description{
The sm has been changed to head(sm) to avoid excessive output
}
\examples{
\dontrun{myclt(n=10,iter=10000)}
}
