% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Density Function}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{p < a}
}
\value{
a plot with the normal density colored in
}
\description{
Normal Density Function
}
\examples{
myncurve(10, 5, 2)
}
